Add script to cache blockchain in the background and load cached initial state
To speed up initial load time.

Run script using:
`npm run cache-arc-dev` which is the same as `NODE_ENV=development ts-node src/bg_cache_worker.ts` and uses an API_URL of https://127.0.0.1:3001
`npm run cache-arc-prod` which is the same as `NODE_ENV=production ts-node src/bg_cache_worker.ts` and uses an API_URL of https://127.0.0.1:3001

By default it tries to cache a local ganache instance, but you can set the environment variable network to change which chain is cached. e.g.
`network=kovan npm run cache-arc-prod`

This creates a .json file in the filesystem which is loaded by the app to pull in the initial state. The file corresponds to the blockchain that was cached so the Koran cache will only be loaded if network=kovan when running the app.