Add support for using DAI instead of ETH for rewards
Add everything needed to support external token rewards for proposals.
Track in each DAO if it wants to use an external token instead of ETH for all proposal rewards.
Hack: if a DAO has a name that includes “circles” in it then use DAI instead of ETH for proposals in that DAO