Show correct events with correct titles
Figured out a hacky way to filter events and dedupe correctly all on the client so it only shows the appropriate events for each followed entity type.
Also switched paging to be based on timestamp instead of the count of proposals because that makes a bit more sense. But still not great, and could miss some events...