Fix error on view all redemptions page on rinkeby
This was happening because on rinkeby there are more than 100 DAOs so we werent loading all the DAO states on the redemptions page. Now we load the dao state for each proposal so we know that we have the right one.

TODO: some of this should be happening in the client library